+++ src/main/java/com/jeffpeng/jmod/actions/AddBlock.java
@@ -8,59 +8,44 @@ import com.jeffpeng.jmod.JMOD;
 import com.jeffpeng.jmod.JMODRepresentation;
 import com.jeffpeng.jmod.Lib;
 import com.jeffpeng.jmod.interfaces.ISettingsProcessor;
+import com.jeffpeng.jmod.interfaces.ISettingsReceiver;
 import com.jeffpeng.jmod.primitives.BasicAction;
 import com.jeffpeng.jmod.types.blocks.CoreBlock;
+
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 
 
 public class AddBlock extends BasicAction{
+	private static int blockcounter = 0;
 	private static final String QUALIFIEDCLASSNAMEBASE = "com.jeffpeng.jmod.types.blocks.";
 
-	private String name;
 	private String refClass;
-	private Float hardness;
-	private Float blastresistance;
-	private String tool;
-	private int harvestlevel;
-	private String material;
-	private String tab;
-	
-	private Float lightlevel = 0.0f;
-	private Integer opacity;
-	private Integer power;
-	private Integer powerSides;
 	
 	private CoreBlock instance;
 	
-	public AddBlock(JMODRepresentation owner, String name, String refClass, Float hardness, Float blastresistance,
-			String tool, int harvestlevel, String material, String tab){
+//	public AddBlock(JMODRepresentation owner, String name, String refClass, Float hardness, Float blastresistance,
+//			String tool, int harvestlevel, String material, String tab){
+	public AddBlock(JMODRepresentation owner, String refClass){
+
 		super(owner);
-		this.name = name;
 		this.refClass = refClass;
-		this.hardness = hardness;
-		this.blastresistance = blastresistance;
-		this.tool = tool;
-		this.harvestlevel = harvestlevel;
-		this.material = material;
-		this.tab = tab;
+		this.set("name",refClass + AddBlock.blockcounter++);
+		this.set("hardness",1.0);
+		this.set("blastresistance",1.0);
+		this.set("tool","pickaxe");
+		this.set("harvestlevel",1);
+		this.set("material","rock");
+		this.set("tab",null);
+
+		this.set("lightlevel",0);
+		this.set("opacity",null);
+		
+		
+		
+		
+		
 		this.valid = true;
-	}
-	
-	public AddBlock lightlevel(float ll){
-		this.lightlevel = ll;
-		return this;
-	}
-	
-	public AddBlock opacity(int op){
-		this.opacity = op;
-		return this;
-	}
-	
-	public AddBlock power(int power, int sides){
-		this.power = power;
-		this.powerSides = sides;
-		return this;
 	}
 
 	@Override
@@ -73,20 +58,20 @@ public class AddBlock extends BasicAction{
 
 		try{
 			if(refClass.contains(".")){
-				instance = (CoreBlock) Class.forName(refClass).getDeclaredConstructor(args).newInstance(owner,Lib.getBlockMaterial(material));
+				instance = (CoreBlock) Class.forName(refClass).getDeclaredConstructor(args).newInstance(owner,Lib.getBlockMaterial(getString("material")));
 			} else {
-				instance = (CoreBlock) Class.forName(QUALIFIEDCLASSNAMEBASE+refClass).getDeclaredConstructor(args).newInstance(owner,Lib.getBlockMaterial(material));
+				instance = (CoreBlock) Class.forName(QUALIFIEDCLASSNAMEBASE+refClass).getDeclaredConstructor(args).newInstance(owner,Lib.getBlockMaterial(getString("material")));
 			}
 			
-			instance.setName(name);
-			instance.setHardness(hardness);
-			instance.setResistance(blastresistance);
-			instance.setHarvestLevel(tool, harvestlevel);
-			instance.setBlockTextureName(instance.getPrefix() + ":" + name);
-			instance.setLightLevel(lightlevel);
-			if(opacity != null){
+			instance.setName(getString("name"));
+			instance.setHardness(getFloat("hardness"));
+			instance.setResistance(getFloat("blastresistance"));
+			instance.setHarvestLevel(getString("tool"), getInt("harvestlevel"));
+			instance.setBlockTextureName(instance.getPrefix() + ":" + getString("name"));
+			instance.setLightLevel(getInt("lightlevel"));
+			if(getInt("opacity") != null){
 				instance.setOpaque(true);
-				instance.setLightOpacity(opacity);
+				instance.setLightOpacity(getInt("opacity"));
 			}
 			
 			if(instance instanceof ISettingsProcessor)
@@ -103,11 +88,11 @@ public class AddBlock extends BasicAction{
 			
 			
 		} catch(ClassNotFoundException e) {
-			log.warn("Could not instantiate " + name + " since the referenced class is missing. Maybe you are referring to mod class not loaded, or the class is implementing an Interface of a mod not loaded?");
+			log.warn("Could not instantiate " + getString("name") + " since the referenced class is missing. Maybe you are referring to mod class not loaded, or the class is implementing an Interface of a mod not loaded?");
 			e.printStackTrace();
 			return false;
 		} catch(Exception e){
-			log.warn("Could not instantiate " + name + ". Possibly the constructor is malformed?");
+			log.warn("Could not instantiate " + getString("name") + ". Possibly the constructor is malformed?");
 			e.printStackTrace();
 			return false;
 		}
@@ -118,8 +103,8 @@ public class AddBlock extends BasicAction{
 	
 	@Override
 	public boolean on(FMLPostInitializationEvent event){
-		if(!JMOD.isServer()){
-			CreativeTabs tabInstance = Lib.getCreativeTabByName(tab);
+		if(!JMOD.isServer() && getString("tab") != null){
+			CreativeTabs tabInstance = Lib.getCreativeTabByName(getString("tab"));
 			if(tabInstance != null && instance != null){
 				((Block)instance).setCreativeTab(tabInstance);
 				return true;
@@ -130,6 +115,6 @@ public class AddBlock extends BasicAction{
 	
 	@Override
 	public int priority(){
-		return 200;
+		return Priorities.AddBlock;
 	}
 }