+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 
-beta1.2.1 (LTS release)
+alpha1.3.1
+- Implemented Basic and "Dynamic" blocks
+- Implemented support for RotaryCraft SurrogateBedrock on Dynamic Blocks
+- Implemented support for ReactorCrafy NeutronShield on Dynamic Blocks
+- Implemented support for Redstone Power on Dynamic Blocks
+- Implemented first step of the EventSystem
+- Introduction of ISettingsReceiver and ISettingsProcessor, part of the syntax rewrite
+- Fixed the Anvil handler AGAIN [Issue #41](https://github.com/SvenKayser/JMOD/issues/41)
+
+
+beta1.2.1 (no release)
 - Implemented proper support for ingredient based search for Shaped StringRecipes for NEI [Issue #31](https://github.com/SvenKayser/JMOD/issues/31)
 - fixed RoC blast furnace handler not executing [Issue #32](https://github.com/SvenKayser/JMOD/issues/32)